pyproject.toml
[project]
name = "llm-history"
version = "0.1.0"
description = "Interactive web app explaining the evolution of language models"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "gensim==4.3.3",
    "numpy==1.26.4",
    "scikit-learn==1.5.2",
    "jinja2==3.1.4",
    "transformers>=4.51,<5",
    "accelerate>=1.0",
    "python-multipart==0.0.9",
    "bitsandbytes>=0.43",
    "tiktoken>=0.7",
    "protobuf>=4.25",
    "sentencepiece>=0.2",
    "torch>=2.12.0.dev0",
    "triton>=3.6.0.dev0",
    "umap-learn>=0.5.7",
    "pillow>=12.2.0",
    "openai>=2.31.0",
    "fal-client>=0.13.2",
]

[project.optional-dependencies]
test = [
    "playwright==1.47.0",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
]

[[tool.uv.index]]
name = "pytorch-nightly"
url = "https://download.pytorch.org/whl/nightly/cu128"
explicit = true

[tool.uv]
prerelease = "if-necessary-or-explicit"

[tool.uv.sources]
torch = { index = "pytorch-nightly" }
triton = { index = "pytorch-nightly" }

[dependency-groups]
dev = [
    "playwright>=1.47.0",
]
